Implement non-binary class probability ranking
Oriented on GoldenEye++, the implementation computes the correlation between the prediction probabilities of the original and permuted data set. In contrary to GoldenEye++, this implementation is not restricted to binary classification.

The current implementation is not usable with upsampling the data to ensure accuracy, as the distance matrix computation would take too long.